Replace expect.assertions example with working one
The current async example in the docs makes no sense, as the test will simply finish before the callbacks are run. As far as I can tell there's no version of `doAsync` that would make that example work correctly without mocking internals of it, or using fake timers if it uses `setTimeout`.